requirements.txt
opencv-python
Pillow
openslide-python
pyvips
pyradiomics
numpy
matplotlib
pymongo
